+++ swift/codegen/generators/trapgen.py
@@ -18,6 +18,7 @@ import inflection
 from toposort import toposort_flatten
 
 from swift.codegen.lib import dbscheme, cpp
+from swift.codegen.loaders import dbschemeloader
 
 log = logging.getLogger(__name__)
 
@@ -73,7 +74,7 @@ def generate(opts, renderer):
     out = opts.cpp_output
 
     traps = {pathlib.Path(): []}
-    for e in dbscheme.iterload(opts.dbscheme):
+    for e in dbschemeloader.iterload(opts.dbscheme):
         if e.is_table:
             traps.setdefault(e.dir, []).append(get_trap(e))
         elif e.is_union:
