+++ swift/codegen/lib/BUILD.bazel
@@ -2,11 +2,10 @@ load("@swift_codegen_deps//:requirements.bzl", "requirement")
 
 py_library(
     name = "lib",
-    srcs = glob(["**/*.py"]),
+    srcs = glob(["*.py"]),
     visibility = ["//swift/codegen:__subpackages__"],
     deps = [
         requirement("pystache"),
-        requirement("pyyaml"),
         requirement("inflection"),
     ],
 )
