--- swift/codegen/lib/schema/__init__.py
@@ -1,1 +1,0 @@
-from .schema import *
