+++ swift/codegen/loaders/BUILD.bazel
@@ -1,11 +1,11 @@
 load("@swift_codegen_deps//:requirements.bzl", "requirement")
 
 py_library(
-    name = "generators",
+    name = "loaders",
     srcs = glob(["*.py"]),
     visibility = ["//swift/codegen:__subpackages__"],
     deps = [
-        "//swift/codegen/lib",
-        "//swift/codegen/loaders",
+        requirement("toposort"),
+        requirement("inflection"),
     ],
 )
