+++ swift/codegen/test/utils.py
@@ -47,7 +47,7 @@ def override_paths(tmp_path):
 @pytest.fixture
 def input(opts, tmp_path):
     opts.schema = tmp_path / schema_file
-    with mock.patch("swift.codegen.lib.schema.load_file") as load_mock:
+    with mock.patch("swift.codegen.loaders.schemaloader.load_file") as load_mock:
         load_mock.return_value = schema.Schema([])
         yield load_mock.return_value
     assert load_mock.mock_calls == [
@@ -58,7 +58,7 @@ def input(opts, tmp_path):
 @pytest.fixture
 def dbscheme_input(opts, tmp_path):
     opts.dbscheme = tmp_path / dbscheme_file
-    with mock.patch("swift.codegen.lib.dbscheme.iterload") as load_mock:
+    with mock.patch("swift.codegen.loaders.dbschemeloader.iterload") as load_mock:
         load_mock.entities = []
         load_mock.side_effect = lambda _: load_mock.entities
         yield load_mock
